requirements.txt
flask 
flask_session
requests
flask_sqlalchemy
tempfile
werkzeug.security
flask-migrate